Refactor key manager into modular components with separate files
